feat: add a11y to roomItem
