+++ app/containers/RoomItem/RoomItem.tsx
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useMemo } from 'react';
 import { View } from 'react-native';
 
 import styles from './styles';
@@ -13,6 +13,7 @@ import Tag from './Tag';
 import I18n from '../../i18n';
 import { DisplayMode } from '../../lib/constants';
 import { IRoomItemProps } from './interfaces';
+import { formatLastMessage } from '../../lib/methods/formatLastMessage';
 
 const RoomItem = ({
 	rid,
@@ -30,7 +31,6 @@ const RoomItem = ({
 	isGroupChat,
 	isRead,
 	date,
-	accessibilityLabel,
 	favorite,
 	lastMessage,
 	alert,
@@ -54,107 +54,115 @@ const RoomItem = ({
 	displayMode,
 	sourceType,
 	hideMentionStatus
-}: IRoomItemProps) => (
-	<Touchable
-		onPress={onPress}
-		onLongPress={onLongPress}
-		width={width}
-		favorite={favorite}
-		toggleFav={toggleFav}
-		isRead={isRead}
-		rid={rid}
-		toggleRead={toggleRead}
-		hideChannel={hideChannel}
-		testID={testID}
-		type={type}
-		isFocused={!!isFocused}
-		swipeEnabled={swipeEnabled}
-		displayMode={displayMode}>
-		<Wrapper
-			accessibilityLabel={accessibilityLabel}
-			avatar={avatar}
-			type={type}
-			userId={userId}
+}: IRoomItemProps) => {
+	// later update this
+	const memoizedMessage = useMemo(
+		() => formatLastMessage({ lastMessage, username, useRealName, showLastMessage, alert, type }),
+		[lastMessage, username, useRealName, showLastMessage, alert, type]
+	);
+	const accessibilityLabel = useMemo(() => `${name} ${memoizedMessage}`, [name, memoizedMessage]);
+	return (
+		<Touchable
+			onPress={onPress}
+			onLongPress={onLongPress}
+			width={width}
+			favorite={favorite}
+			toggleFav={toggleFav}
+			isRead={isRead}
 			rid={rid}
-			prid={prid}
-			status={status}
-			isGroupChat={isGroupChat}
-			teamMain={teamMain}
-			displayMode={displayMode}
-			showAvatar={showAvatar}
-			showLastMessage={!!showLastMessage}
-			sourceType={sourceType}>
-			{showLastMessage && displayMode === DisplayMode.Expanded ? (
-				<>
-					<View style={styles.titleContainer}>
-						{showAvatar ? (
-							<TypeIcon
-								userId={userId}
+			toggleRead={toggleRead}
+			hideChannel={hideChannel}
+			testID={testID}
+			type={type}
+			isFocused={!!isFocused}
+			swipeEnabled={swipeEnabled}
+			displayMode={displayMode}>
+			<Wrapper
+				accessibilityLabel={accessibilityLabel}
+				avatar={avatar}
+				type={type}
+				userId={userId}
+				rid={rid}
+				prid={prid}
+				status={status}
+				isGroupChat={isGroupChat}
+				teamMain={teamMain}
+				displayMode={displayMode}
+				showAvatar={showAvatar}
+				showLastMessage={!!showLastMessage}
+				sourceType={sourceType}>
+				{showLastMessage && displayMode === DisplayMode.Expanded ? (
+					<>
+						<View style={styles.titleContainer}>
+							{showAvatar ? (
+								<TypeIcon
+									userId={userId}
+									type={type}
+									prid={prid}
+									status={status}
+									isGroupChat={isGroupChat}
+									teamMain={teamMain}
+									sourceType={sourceType}
+								/>
+							) : null}
+							<Title name={name} hideUnreadStatus={hideUnreadStatus} alert={alert} />
+							{autoJoin ? <Tag testID='auto-join-tag' name={I18n.t('Auto-join')} /> : null}
+							<UpdatedAt date={date} hideUnreadStatus={hideUnreadStatus} alert={alert} />
+						</View>
+						<View style={styles.row} testID='room-item-last-message-container'>
+							<LastMessage
+								lastMessage={lastMessage}
 								type={type}
-								prid={prid}
-								status={status}
-								isGroupChat={isGroupChat}
-								teamMain={teamMain}
-								sourceType={sourceType}
+								showLastMessage={showLastMessage}
+								username={username || ''}
+								alert={alert && !hideUnreadStatus}
+								useRealName={useRealName}
 							/>
-						) : null}
-						<Title name={name} hideUnreadStatus={hideUnreadStatus} alert={alert} />
-						{autoJoin ? <Tag testID='auto-join-tag' name={I18n.t('Auto-join')} /> : null}
-						<UpdatedAt date={date} hideUnreadStatus={hideUnreadStatus} alert={alert} />
-					</View>
-					<View style={styles.row} testID='room-item-last-message-container'>
-						<LastMessage
-							lastMessage={lastMessage}
+							<UnreadBadge
+								unread={unread}
+								userMentions={userMentions}
+								groupMentions={groupMentions}
+								tunread={tunread}
+								tunreadUser={tunreadUser}
+								tunreadGroup={tunreadGroup}
+								hideMentionStatus={hideMentionStatus}
+								hideUnreadStatus={hideUnreadStatus}
+							/>
+						</View>
+					</>
+				) : (
+					<View style={[styles.titleContainer, styles.flex]}>
+						<TypeIcon
+							userId={userId}
 							type={type}
-							showLastMessage={showLastMessage}
-							username={username || ''}
-							alert={alert && !hideUnreadStatus}
-							useRealName={useRealName}
+							prid={prid}
+							status={status}
+							isGroupChat={isGroupChat}
+							teamMain={teamMain}
+							size={22}
+							style={{ marginRight: 8 }}
+							sourceType={sourceType}
 						/>
-						<UnreadBadge
-							unread={unread}
-							userMentions={userMentions}
-							groupMentions={groupMentions}
-							tunread={tunread}
-							tunreadUser={tunreadUser}
-							tunreadGroup={tunreadGroup}
-							hideMentionStatus={hideMentionStatus}
-							hideUnreadStatus={hideUnreadStatus}
-						/>
+						<Title name={name} hideUnreadStatus={hideUnreadStatus} alert={alert} />
+						{autoJoin ? <Tag name={I18n.t('Auto-join')} /> : null}
+						<View style={styles.wrapUpdatedAndBadge}>
+							<UpdatedAt date={date} hideUnreadStatus={hideUnreadStatus} alert={alert} />
+							<UnreadBadge
+								unread={unread}
+								userMentions={userMentions}
+								groupMentions={groupMentions}
+								tunread={tunread}
+								tunreadUser={tunreadUser}
+								tunreadGroup={tunreadGroup}
+								hideMentionStatus={hideMentionStatus}
+								hideUnreadStatus={hideUnreadStatus}
+							/>
+						</View>
 					</View>
-				</>
-			) : (
-				<View style={[styles.titleContainer, styles.flex]}>
-					<TypeIcon
-						userId={userId}
-						type={type}
-						prid={prid}
-						status={status}
-						isGroupChat={isGroupChat}
-						teamMain={teamMain}
-						size={22}
-						style={{ marginRight: 8 }}
-						sourceType={sourceType}
-					/>
-					<Title name={name} hideUnreadStatus={hideUnreadStatus} alert={alert} />
-					{autoJoin ? <Tag name={I18n.t('Auto-join')} /> : null}
-					<View style={styles.wrapUpdatedAndBadge}>
-						<UpdatedAt date={date} hideUnreadStatus={hideUnreadStatus} alert={alert} />
-						<UnreadBadge
-							unread={unread}
-							userMentions={userMentions}
-							groupMentions={groupMentions}
-							tunread={tunread}
-							tunreadUser={tunreadUser}
-							tunreadGroup={tunreadGroup}
-							hideMentionStatus={hideMentionStatus}
-							hideUnreadStatus={hideUnreadStatus}
-						/>
-					</View>
-				</View>
-			)}
-		</Wrapper>
-	</Touchable>
-);
+				)}
+			</Wrapper>
+		</Touchable>
+	);
+};
 
 export default RoomItem;
